Add files via upload
Added stuff for store and fixed the big loop in calc_scoreboard. I still need to add the stuff for the WAR hazard. I plan to add all of the other initial arithmetic for each of the instructions first so we can test it much better. Once that is complete we work on big algorithm.